+++ be15_DevEagles_FE/src/components/common/index.js
@@ -11,7 +11,7 @@ export { default as BaseTable } from './BaseTable.vue';
 export { default as BaseLoading } from './BaseLoading.vue';
 export { default as BaseToast } from './BaseToast.vue';
 export { default as BaseDrawer } from './BaseDrawer.vue';
-export { default as BasePagination } from './Pagaination.vue';
+export { default as Pagination } from './Pagination.vue';
 export { default as BaseCalendar } from './BaseCalendar.vue';
 export { default as VCalendarWrapper } from './VCalendarWrapper.vue';
