+++ be15_DevEagles_FE/src/features/coupons/views/CouponList.vue
@@ -117,7 +117,7 @@
   import BaseButton from '@/components/common/BaseButton.vue';
   import BaseModal from '@/components/common/BaseModal.vue';
   import BasePopover from '@/components/common/BasePopover.vue';
-  import BasePagination from '@/components/common/Pagaination.vue';
+  import Pagination from '@/components/common/Pagination.vue';
   import BaseCard from '@/components/common/BaseCard.vue';
   import BaseTable from '@/components/common/BaseTable.vue';
   import BaseBadge from '@/components/common/BaseBadge.vue';
@@ -132,7 +132,7 @@
       BaseButton,
       BaseModal,
       BasePopover,
-      BasePagination,
+      BasePagination: Pagination,
       BaseCard,
       BaseTable,
       BaseBadge,
